Export call logs using Companion App, not adb

+++ functions/backup_func.sh
@@ -55,24 +55,33 @@ function backup_func() {
   done
 
   # Export contacts and SMS messages
-  cecho "Exporting contacts (as vCard) and SMS messages (as CSV)."
+  cecho "Exporting contacts (as vCard), call logs as well as SMS messages (as CSV)."
+  # Get the entire oab-temp directory
+  mkdir ./backup-tmp/open-android-backup-temp
+  if ! get_file /storage/emulated/0/open-android-backup-temp . ./backup-tmp/open-android-backup-temp; then
+    cecho "Error: Failed to get data from the Companion App! Please make sure that you have pressed the 'Export Data' button in the Companion App."
+    cecho "If you have already done that, please report this issue on GitHub."
+    cecho "Cannot continue - exiting."
+    exit 1
+  fi
+  # Get contacts
   mkdir ./backup-tmp/Contacts
-  get_file /storage/emulated/0/open-android-backup-temp . ./backup-tmp/Contacts
+  mv ./backup-tmp/open-android-backup-temp/open-android-backup-contact*.vcf ./backup-tmp/Contacts || cecho "No contacts found on device - ignoring." 1>&2
+  # Get SMS messages
   mkdir ./backup-tmp/SMS
-  mv ./backup-tmp/Contacts/SMS_Messages.csv ./backup-tmp/SMS
+  mv ./backup-tmp/open-android-backup-temp/SMS_Messages.csv ./backup-tmp/SMS
+  # Get call logs
+  mkdir ./backup-tmp/CallLogs
+  mv ./backup-tmp/open-android-backup-temp/Call_Logs.csv ./backup-tmp/CallLogs
+  # Cleanup
   cecho "Removing temporary files created by the companion app."
   adb shell rm -rf /storage/emulated/0/open-android-backup-temp
+  rm -rf ./backup-tmp/open-android-backup-temp
 
   # Export internal storage
   cecho "Exporting internal storage - this will take a while."
   mkdir ./backup-tmp/Storage
   get_file /storage/emulated/0 . ./backup-tmp/Storage
-
-  # Export call logs
-  cecho "Exporting call logs."
-  mkdir ./backup-tmp/CallLogs
-  adb shell content query --uri content://call_log/calls --projection name:normalized_number:duration:via_number:geocoded_location:date > ./backup-tmp/CallLogs/calls.txt || cecho "Couldn't backup call logs due to an error - ignoring." 1>&2
-  adb shell content query --uri content://call_log/calls > ./backup-tmp/CallLogs/raw_data.txt || cecho "Couldn't backup raw call logs due to an error - ignoring." 1>&2
 
   # Run the third-party backup hook, if enabled.
   if [ "$use_hooks" = "yes" ] && [ "$(type -t backup_hook)" == "function" ]; then
